Add pm1 merging
